+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 * NP-214 Link benefits to the fully qualified URL of the public website.
+* NP-239 Advice Feedback
 * NP-301 Textarea control
 
 ## <sub>v0.9.0</sub>